Code refactor : Add PrefHelper and DatabaseHelper
