--- debug-db/src/main/java/com/amitshekhar/utils/PrefUtils.java
@@ -1,60 +1,0 @@
-/*
- *
- *  *    Copyright (C) 2016 Amit Shekhar
- *  *    Copyright (C) 2011 Android Open Source Project
- *  *
- *  *    Licensed under the Apache License, Version 2.0 (the "License");
- *  *    you may not use this file except in compliance with the License.
- *  *    You may obtain a copy of the License at
- *  *
- *  *        http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  *    Unless required by applicable law or agreed to in writing, software
- *  *    distributed under the License is distributed on an "AS IS" BASIS,
- *  *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  *    See the License for the specific language governing permissions and
- *  *    limitations under the License.
- *
- */
-
-package com.amitshekhar.utils;
-
-import android.content.Context;
-
-import java.io.File;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-/**
- * Created by amitshekhar on 16/11/16.
- */
-
-public final class PrefUtils {
-
-    private static final String PREFS_SUFFIX = ".xml";
-
-    private PrefUtils() {
-        // This class in not publicly instantiable
-    }
-
-    public static List<String> getSharedPreferenceTags(Context context) {
-
-        ArrayList<String> tags = new ArrayList<>();
-
-        String rootPath = context.getApplicationInfo().dataDir + "/shared_prefs";
-        File root = new File(rootPath);
-        if (root.exists()) {
-            for (File file : root.listFiles()) {
-                String fileName = file.getName();
-                if (fileName.endsWith(PREFS_SUFFIX)) {
-                    tags.add(fileName.substring(0, fileName.length() - PREFS_SUFFIX.length()));
-                }
-            }
-        }
-
-        Collections.sort(tags);
-
-        return tags;
-    }
-}
